owkin-k/requirements.txt
# Gateway
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MCP Servers
fastmcp>=2.0.0

# Agent Runtime
anthropic>=0.40.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# CLI
click>=8.1.0
rich>=13.0.0
cookiecutter>=2.5.0

# Observability
opentelemetry-sdk>=1.28.0
